+++ ICanPay/ICanPay.Core/ICanPay.Core.csproj
@@ -11,12 +11,17 @@
     <DebugSymbols>false</DebugSymbols>
   </PropertyGroup>
 
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <PackageReference Include="Microsoft.AspNetCore.Http">
+      <Version>2.0.0</Version>
+    </PackageReference>
+  </ItemGroup>
+
   <ItemGroup Condition="'$(TargetFramework)' == 'net35'">
     <Reference Include="System.Web" />
+    <PackageReference Include="ThoughtWorks.QRCode">
+      <Version>1.1.0</Version>
+    </PackageReference>
   </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ICanPay.Core\ICanPay.Core.csproj" />
-  </ItemGroup>
-
+  
 </Project>
