+++ ICanPay/ICanPay.Core/StaticHttpContextExtensions.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 #endif
 
-namespace ICanPay
+namespace ICanPay.Core
 {
 #if NETSTANDARD2_0
