+++ ICanPay/ICanPay.Core/UnknownGatewayEventArgs.cs
@@ -1,0 +1,24 @@
+﻿
+namespace ICanPay.Core
+{
+    /// <summary>
+    /// 未知网关事件数据
+    /// </summary>
+    public class UnknownGatewayEventArgs : PaymentEventArgs
+    {
+
+        #region 构造函数
+
+        /// <summary>
+        /// 初始化未知网关事件数据
+        /// </summary>
+        /// <param name="gateway">支付网关</param>
+        public UnknownGatewayEventArgs(GatewayBase gateway)
+            : base(gateway)
+        {
+        }
+
+        #endregion
+
+    }
+}
