--- ICanPay/ICanPay/IPaymentQRCode.cs
@@ -1,16 +1,0 @@
-﻿
-namespace ICanPay
-{
-    /// <summary>
-    /// 订单是使用二维码支付时创建订单的支付二维码
-    /// </summary>
-    internal interface IPaymentQRCode
-    {
-
-        /// <summary>
-        /// 获得订单的支付二维码内容
-        /// </summary>
-        /// <returns>订单的支付二维码内容</returns>
-        string GetPaymentQRCodeContent();
-    }
-}
